# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Sistema de Gestión Bancaria
  description: API para gestionar productos bancarios.
  version: 1.0.0
tags:
  - name: products
    description: All about your product information
    externalDocs:
      description: Find out more
      url: http://swagger.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /products/{customId}:
    get:
      tags:
        - products
      summary: Get list of all products
      parameters:
        - name: customId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Client List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '500':
          description: Server error

  /products:
    get:
      tags:
        - products
      summary: Get list of all products
      responses:
        '200':
          description: Client List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductResponse'
        '500':
          description: Server error

    post:
      tags:
        - products
      summary: Create a new product
      requestBody:
        description: Data to create the product
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '201':
          description: Product created


  /products/{id}:
    get:
      tags:
        - products
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'

    put:
      tags:
        - products
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        description: Data to update the product
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductRequest'
      responses:
        '202':
          description: Updated client

        '404':
          description: Product not found

    delete:
      tags:
        - products
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted product


components:
  schemas:
    ProductResponse:
      type: object
      properties:
        id:
          type: string
          description: unique id
        productType:
          type: string
          description: Tipo de producto (AHORRO, CORRIENTE, PLAZO_FIJO)

        balance:
          type: number
          format: double
          description: Saldo
        limitMnthlyMovements:
          type: integer
          description: Límite de movimientos mensuales para cuentas de ahorro
        dayMovement:
          type: string
          description: Día de movimiento para cuentas a plazo fijo
        customerId:
          type: string
          description: Id cliente
        movementId:
          type: string
          description: Id movimiento

    ProductRequest:
      type: object
      properties:
        productType:
          type: string
          description: |
            Tipo de producto financiero ofrecido por el banco. Cada tipo incluye una categoría y una descripción:
            - **Pasivo**: Productos como cuentas bancarias.
            - **Activo**: Productos como créditos y tarjetas.
          enum:
            - AHORRO
            - CUENTA_CORRIENTE
            - PLAZO_FIJO
            - CREDITO_PERSONAL
            - CREDITO_EMPRESARIAL
            - TARJETA_CREDITO

        balance:
          type: number
          format: double
          description: Saldo
        limitMnthlyMovements:
          type: integer
          description: Límite de movimientos mensuales para cuentas de ahorro
        dayMovement:
          type: string
          description: Día de movimiento para cuentas a plazo fijo
        clientId:
          type: string
          description: Id client
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Code Error
        message:
          type: string
          description: Message